+++ src/LiveCharts.Core/Abstractions/DataSeries/IScatterSeries.cs
@@ -22,7 +22,7 @@ namespace LiveCharts.Core.Abstractions.DataSeries
         /// <value>
         /// The maximum point diameter.
         /// </value>
-        double MaxPointDiameter { get; set; }
+        float MaxPointDiameter { get; set; }
 
         /// <summary>
         /// Gets or sets the minimum point diameter.
@@ -30,6 +30,6 @@ namespace LiveCharts.Core.Abstractions.DataSeries
         /// <value>
         /// The minimum point diameter.
         /// </value>
-        double MinPointDiameter { get; set; }
+        float MinPointDiameter { get; set; }
     }
 }
